fix: add Epic vendor quirk to pass Epic-Client-ID
